database_management/requirements.txt
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
geoalchemy2>=0.14.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
click>=8.0.0
rich>=13.0.0
pathlib
typing-extensions>=4.0.0